WebCrawler-master/requirements.txt
pymongo
selenium
webdriver_manager
psycopg2-binary
scrapy-user-agents